+++ adapters/DTExchange/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #### Next version
 - Added didFailToPresentWithError callbacks in rewarded ad.
+- Added IAAdDidExpire delegate method in interstitial and rewarded ads.
 
 #### [Version 8.2.6.0](https://dl.google.com/googleadmobadssdk/mediation/ios/dtexchange/DTExchangeAdapter-8.2.6.0.zip)
 - Verified compatibility with DT Exchange SDK version 8.2.6.